requirements.txt
redis
requests
diagramsf